debezium/dbz#2235 Map MySQL and MariaDB signed TINYINT columns to INT8
MySQL and MariaDB signed TINYINT is an 8-bit type (-128 to 127), yet
the ANTLR DDL parsers folded the TINYINT and INT1 keywords into
java.sql.Types.SMALLINT, so signed TINYINT columns were emitted with an
INT16 schema. This was the only integer mapping that did not use the
smallest Kafka Connect type capable of holding the source range;
Types.TINYINT already mapped to INT8 in the shared JDBC converters, but
no binlog column ever reached that path.

Register TINYINT and INT1 as Types.TINYINT in all three parsers so that
signed TINYINT columns take the core Types.TINYINT path and are emitted
as INT8 with Byte values. Unsigned handling keeps the established
widen-by-one-type rule via the type-name based overrides: a new
explicit rule emits TINYINT UNSIGNED (0 to 255) as INT16, alongside an
explicit MEDIUMINT UNSIGNED rule that documents the existing INT32
result now that the jdbcType no longer decides these cases.

Align convertTinyInt with the INT8 schema by producing Byte values
(previously it delegated to convertSmallInt and produced Short), and
extend the decimal-string default value handling to the tinyint
converter so defaults such as '10.01' keep working. Default values for
signed TINYINT columns are now emitted as Byte accordingly.

+++ debezium-connector-binlog/src/test/java/io/debezium/connector/binlog/BinlogAntlrDdlParserTest.java
@@ -681,7 +681,7 @@ public abstract class BinlogAntlrDdlParserTest<V extends BinlogValueConverters, 
         assertThat(table.columnWithName("mi").jdbcType()).isEqualTo(Types.INTEGER);
         assertThat(table.columnWithName("f4").jdbcType()).isEqualTo(Types.FLOAT);
         assertThat(table.columnWithName("f8").jdbcType()).isEqualTo(Types.DOUBLE);
-        assertThat(table.columnWithName("i1").jdbcType()).isEqualTo(Types.SMALLINT);
+        assertThat(table.columnWithName("i1").jdbcType()).isEqualTo(Types.TINYINT);
         assertThat(table.columnWithName("i2").jdbcType()).isEqualTo(Types.SMALLINT);
         assertThat(table.columnWithName("i3").jdbcType()).isEqualTo(Types.INTEGER);
         assertThat(table.columnWithName("i4").jdbcType()).isEqualTo(Types.INTEGER);
@@ -2208,7 +2208,7 @@ public abstract class BinlogAntlrDdlParserTest<V extends BinlogValueConverters, 
         assertThat(t).isNotNull();
         assertThat(t.retrieveColumnNames()).containsExactly("action", "revision", "changed_on", "id", "name");
         assertThat(t.primaryKeyColumnNames()).containsExactly("id", "revision");
-        assertColumn(t, "action", "TINYINT UNSIGNED", Types.SMALLINT, 3, -1, false, false, false);
+        assertColumn(t, "action", "TINYINT UNSIGNED", Types.TINYINT, 3, -1, false, false, false);
         assertColumn(t, "revision", "INT UNSIGNED", Types.INTEGER, 10, -1, false, false, false);
         assertColumn(t, "changed_on", "DATETIME", Types.TIMESTAMP, -1, -1, false, false, false);
         assertColumn(t, "id", "VARCHAR", Types.VARCHAR, 36, -1, false, false, false);
@@ -3192,7 +3192,7 @@ public abstract class BinlogAntlrDdlParserTest<V extends BinlogValueConverters, 
         assertThat(table.columnWithName("id").isOptional()).isEqualTo(false);
         assertThat(getColumnSchema(table, "boolean_c").defaultValue()).isEqualTo(false);
         assertThat(getColumnSchema(table, "bit_c").defaultValue()).isEqualTo(true);
-        assertThat(getColumnSchema(table, "tiny_c").defaultValue()).isEqualTo((short) 2);
+        assertThat(getColumnSchema(table, "tiny_c").defaultValue()).isEqualTo((byte) 2);
         assertThat(getColumnSchema(table, "tiny_un_c").defaultValue()).isEqualTo((short) 3);
         assertThat(getColumnSchema(table, "tiny_un_z_c").defaultValue()).isEqualTo((short) 4);
         assertThat(getColumnSchema(table, "small_c").defaultValue()).isEqualTo((short) 5);
